+++ src/dialect/mod.rs
@@ -1,3 +1,4 @@
+pub mod caps;
 pub mod mariadb;
 
 use std::str::FromStr;